@@ -3,13 +3,11 @@ package simsx
 import (
 	"errors"
 	"fmt"
+	"maps"
 	"slices"
-	"sort"
 	"strings"
 	"sync"
 	"sync/atomic"
-
-	"golang.org/x/exp/maps"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
@@ -237,14 +235,18 @@ func (s *ExecutionSummary) Add(module, url string, status ReporterStatus, commen
 func (s *ExecutionSummary) String() string {
 	s.mx.RLock()
 	defer s.mx.RUnlock()
-	keys := maps.Keys(s.counts)
-	sort.Strings(keys)
+	keys := slices.Sorted(maps.Keys(s.counts))
 	var sb strings.Builder
 	for _, key := range keys {
 		sb.WriteString(fmt.Sprintf("%s: %d\n", key, s.counts[key]))
 	}
+	if len(s.skipReasons) != 0 {
+		sb.WriteString("\nSkip reasons:\n")
+	}
 	for m, c := range s.skipReasons {
-		sb.WriteString(fmt.Sprintf("%d\t%s: %q\n", sum(maps.Values(c)), m, maps.Keys(c)))
+		values := maps.Values(c)
+		keys := maps.Keys(c)
+		sb.WriteString(fmt.Sprintf("%d\t%s: %q\n", sum(slices.Collect(values)), m, slices.Collect(keys)))
 	}
 	return sb.String()
 }
