@@ -82,11 +82,11 @@ HbP+c6JmeJy9JXe2rbbF1QtCX1gLqGcDQPBXiCtFvP7/8wTZtVOPj8vREzhZ9ElO
 
 			// Now add a temporary keybase
 			kbHome := filepath.Join(t.TempDir(), fmt.Sprintf("kbhome-%s", tc.name))
-			// Create dir, otherwise os.WriteFile will fail
-			if _, err := os.Stat(kbHome); os.IsNotExist(err) {
-				err = os.MkdirAll(kbHome, 0o700)
-				require.NoError(t, err)
-			}
+			require.NoError(t, os.MkdirAll(kbHome, 0o700))
+			t.Cleanup(func() {
+				require.NoError(t, os.RemoveAll(kbHome))
+			})
+
 			kb, err := keyring.New(sdk.KeyringServiceName(), tc.keyringBackend, kbHome, nil, cdc)
 			require.NoError(t, err)
 
@@ -97,14 +97,8 @@ HbP+c6JmeJy9JXe2rbbF1QtCX1gLqGcDQPBXiCtFvP7/8wTZtVOPj8vREzhZ9ElO
 				WithCodec(cdc)
 			ctx := context.WithValue(context.Background(), client.ClientContextKey, &clientCtx)
 
-			t.Cleanup(cleanupKeys(t, kb, "keyname1"))
-
 			keyfile := filepath.Join(kbHome, "key.asc")
 			require.NoError(t, os.WriteFile(keyfile, []byte(armoredKey), 0o600))
-
-			defer func() {
-				_ = os.RemoveAll(kbHome)
-			}()
 
 			mockIn.Reset(tc.userInput)
 			cmd.SetArgs([]string{
@@ -128,12 +122,20 @@ func Test_runImportHexCmd(t *testing.T) {
 		name           string
 		keyringBackend string
 		hexKey         string
+		stdInput       bool
 		keyType        string
 		expectError    bool
 	}{
 		{
 			name:           "test backend success",
 			keyringBackend: keyring.BackendTest,
+			hexKey:         "0xa3e57952e835ed30eea86a2993ac2a61c03e74f2085b3635bd94aa4d7ae0cfdf",
+			keyType:        "secp256k1",
+		},
+		{
+			name:           "read the hex key from standard input",
+			keyringBackend: keyring.BackendTest,
+			stdInput:       true,
 			hexKey:         "0xa3e57952e835ed30eea86a2993ac2a61c03e74f2085b3635bd94aa4d7ae0cfdf",
 			keyType:        "secp256k1",
 		},
@@ -147,6 +149,10 @@ func Test_runImportHexCmd(t *testing.T) {
 
 			// Now add a temporary keybase
 			kbHome := filepath.Join(t.TempDir(), fmt.Sprintf("kbhome-%s", tc.name))
+			t.Cleanup(func() {
+				require.NoError(t, os.RemoveAll(kbHome))
+			})
+
 			kb, err := keyring.New(sdk.KeyringServiceName(), tc.keyringBackend, kbHome, nil, cdc)
 			require.NoError(t, err)
 
@@ -157,17 +163,17 @@ func Test_runImportHexCmd(t *testing.T) {
 				WithCodec(cdc)
 			ctx := context.WithValue(context.Background(), client.ClientContextKey, &clientCtx)
 
-			t.Cleanup(cleanupKeys(t, kb, "keyname1"))
-
-			defer func() {
-				_ = os.RemoveAll(kbHome)
-			}()
-
-			cmd.SetArgs([]string{
-				"keyname1", tc.hexKey,
+			args := []string{"keyname1"}
+			if tc.stdInput {
+				mockIn.Reset(tc.hexKey)
+			} else {
+				args = append(args, tc.hexKey)
+			}
+			cmd.SetArgs(append(
+				args,
 				fmt.Sprintf("--%s=%s", flags.FlagKeyType, tc.keyType),
 				fmt.Sprintf("--%s=%s", flags.FlagKeyringBackend, tc.keyringBackend),
-			})
+			))
 
 			err = cmd.ExecuteContext(ctx)
 			if tc.expectError {
