@@ -11,8 +11,6 @@ import (
 	abciserver "github.com/cometbft/cometbft/abci/server"
 	cmtcmd "github.com/cometbft/cometbft/cmd/cometbft/commands"
 	cmtcfg "github.com/cometbft/cometbft/config"
-	cmtcrypto "github.com/cometbft/cometbft/crypto"
-	cmted25519 "github.com/cometbft/cometbft/crypto/ed25519"
 	"github.com/cometbft/cometbft/node"
 	"github.com/cometbft/cometbft/p2p"
 	pvm "github.com/cometbft/cometbft/privval"
@@ -104,7 +102,6 @@ func (s *CometBFTServer[T]) Init(appI serverv2.AppI[T], cfg map[string]any, logg
 	consensus := NewConsensus(
 		s.logger,
 		appI.Name(),
-		appI.GetConsensusAuthority(),
 		appI.GetAppManager(),
 		s.serverOptions.Mempool(cfg),
 		indexEvents,
@@ -160,9 +157,7 @@ func (s *CometBFTServer[T]) Start(ctx context.Context) error {
 	pv, err := pvm.LoadOrGenFilePV(
 		s.config.ConfigTomlConfig.PrivValidatorKeyFile(),
 		s.config.ConfigTomlConfig.PrivValidatorStateFile(),
-		func() (cmtcrypto.PrivKey, error) {
-			return cmted25519.GenPrivKey(), nil
-		},
+		s.serverOptions.KeygenF,
 	)
 	if err != nil {
 		return err
