@@ -4,23 +4,24 @@ package testutil
 // Those constants are defined here to be used in the SDK without importing those modules.
 
 const (
-	AccountsModuleName     = "accounts"
-	AuthModuleName         = "auth"
-	AuthzModuleName        = "authz"
-	BankModuleName         = "bank"
-	CircuitModuleName      = "circuit"
-	DistributionModuleName = "distribution"
-	EvidenceModuleName     = "evidence"
-	FeegrantModuleName     = "feegrant"
-	GovModuleName          = "gov"
-	GroupModuleName        = "group"
-	MintModuleName         = "mint"
-	NFTModuleName          = "nft"
-	ParamsModuleName       = "params"
-	ProtocolPoolModuleName = "protocolpool"
-	SlashingModuleName     = "slashing"
-	StakingModuleName      = "staking"
-	TxModuleName           = "tx"
-	UpgradeModuleName      = "upgrade"
-	EpochsModuleName       = "epochs"
+	AccountsModuleName              = "accounts"
+	AuthModuleName                  = "auth"
+	AuthzModuleName                 = "authz"
+	BankModuleName                  = "bank"
+	CircuitModuleName               = "circuit"
+	DistributionModuleName          = "distribution"
+	EvidenceModuleName              = "evidence"
+	FeegrantModuleName              = "feegrant"
+	GovModuleName                   = "gov"
+	GroupModuleName                 = "group"
+	MintModuleName                  = "mint"
+	NFTModuleName                   = "nft"
+	ParamsModuleName                = "params"
+	ProtocolPoolModuleName          = "protocolpool"
+	SlashingModuleName              = "slashing"
+	StakingModuleName               = "staking"
+	AuthTxConfigDepinjectModuleName = "tx"
+	UpgradeModuleName               = "upgrade"
+	EpochsModuleName                = "epochs"
+	ValidateModuleName              = "validate"
 )
